refactor: group items redux files
