+++ src/reducers/index.js
@@ -3,9 +3,9 @@ import { routerReducer } from 'react-router-redux'
 import loading from './loadingReducer'
 import error from './errorReducer'
 import date from '../store/date/dateReducer'
-import items from './itemsReducer'
+import items from '../store/expenses/expensesReducer'
 import authentication from '../store/auth/authReducer'
-import expense from './expenseReducer'
+import expense from '../store/currentExpense/currentExpenseReducer'
 
 export default combineReducers({
   routing: routerReducer,
