+++ src/scenes/addExpense/index.js
@@ -1,6 +1,6 @@
 import { push } from 'react-router-redux'
 import { connect } from 'react-redux'
-import { addItem } from '../../actions/items'
+import { addItem } from '../../store/expenses/expensesActions'
 import AddExpense from './AddExpense'
 import { createLoadingSelector } from '../../selectors/loadingSelectors'
