+++ src/store/expenses/expensesSagas.js
@@ -3,9 +3,6 @@ import {
   FETCH_ITEMS_REQUEST,
   FETCH_ITEMS_ERROR,
   FETCH_ITEMS_SUCCESS,
-  FETCH_ITEM_REQUEST,
-  FETCH_ITEM_ERROR,
-  FETCH_ITEM_SUCCESS,
   ADD_ITEM_SUCCESS,
   ADD_ITEM_ERROR,
   ADD_ITEM_REQUEST,
@@ -15,8 +12,8 @@ import {
   UPDATE_ITEM_REQUEST,
   UPDATE_ITEM_SUCCESS,
   UPDATE_ITEM_ERROR,
-} from '../actions/types'
-import itemsApi from '../api/itemsApi'
+} from './expensesActions'
+import itemsApi from '../../api/itemsApi'
 
 function* fetchItems() {
   try {
@@ -26,18 +23,6 @@ function* fetchItems() {
     yield put({
       type: FETCH_ITEMS_ERROR,
       payload: { message: 'Failed API call to fetch all items.' },
-    })
-  }
-}
-
-function* fetchItem(action) {
-  try {
-    const response = yield call(itemsApi.getOne, action.payload)
-    yield put({ type: FETCH_ITEM_SUCCESS, payload: response })
-  } catch (err) {
-    yield put({
-      type: FETCH_ITEM_ERROR,
-      payload: { message: err },
     })
   }
 }
@@ -84,10 +69,6 @@ function* updateItem({ payload }) {
 // Starts fetchUser on each dispatched `USER_FETCH_REQUESTED` action. Allows concurrent fetches of user.
 export function* fetchItemsSaga() {
   yield takeEvery(FETCH_ITEMS_REQUEST, fetchItems)
-}
-
-export function* fetchItemSaga() {
-  yield takeEvery(FETCH_ITEM_REQUEST, fetchItem)
 }
 
 export function* addItemSaga() {
