+++ src/store/expenses/expensesSelectors.js
@@ -1,7 +1,7 @@
 import { createSelector } from 'reselect'
 import { and, propEq, filter, find, gte, lt, prop, sort } from 'ramda'
 import { differenceInDays } from 'date-fns'
-import { getTotalCost } from '../utilities/calculations'
+import { getTotalCost } from '../../utilities/calculations'
 
 const dateDiff = (a, b) => new Date(prop('date', b)) - new Date(prop('date', a))
