+++ src/store/sagas.js
@@ -1,12 +1,12 @@
 import { all } from 'redux-saga/effects'
 import {
   fetchItemsSaga,
-  fetchItemSaga,
   addItemSaga,
   removeItemSaga,
   updateItemSaga,
-} from './itemsSagas'
-import { currentSessionSaga, signInSaga } from '../store/auth/authSagas'
+} from './expenses/expensesSagas'
+import { fetchItemSaga } from './currentExpense/currentExpenseSagas'
+import { currentSessionSaga, signInSaga } from './auth/authSagas'
 
 export default function* rootSaga() {
   yield all([
